% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constroi_obesid_mun.R
\name{constroi_obesid_mun}
\alias{constroi_obesid_mun}
\title{Constroi os dados de Obesidade Municipal - Indicador  Saúde}
\usage{
constroi_obesid_mun()
}
\value{
Um dataframe
}
\description{
Constroi os dados de População de todas as idades que está em situação de obesidade
de acordo com o Índice de Massa Corporal (IMC) com base no Sisvan/Ministério da Saúde	Saúde
}
\examples{
\dontru{
  data <- constroi_obesid_mun()
}
}
