% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.download_file}
\alias{.download_file}
\title{Download a file from a given URL}
\usage{
.download_file(url, dir_path = "data_raw/")
}
\arguments{
\item{url}{Character. The URL of the file to download.}

\item{dir_path}{Character. The directory where the file should be saved. Default is "data_raw/".}
}
\value{
Character. The path to the downloaded file, or NULL if the download fails.
}
\description{
Download a file from a given URL
}
\examples{
file_path <- .download_file("https://example.com/data.zip")
}
