% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.remove_zip}
\alias{.remove_zip}
\title{Remove a ZIP file after extraction}
\usage{
.remove_zip(file_path)
}
\arguments{
\item{file_path}{Character. The path to the ZIP file to be removed.}
}
\value{
NULL. The function removes the ZIP file if it exists.
}
\description{
Remove a ZIP file after extraction
}
\examples{
.remove_zip("data_raw/data.zip")
}
