% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_abandono_escolar.R
\name{download_abandono_escolar}
\alias{download_abandono_escolar}
\title{Download and Extract School Dropout Data}
\usage{
download_abandono_escolar(dir_path = "data_raw/")
}
\arguments{
\item{dir_path}{A character string specifying the directory path where the file will be downloaded and extracted. Default is 'data_raw/'.}
}
\description{
This function downloads and extracts school dropout data from the INEP database.
}
\details{
The function ensures the target directory exists, downloads the ZIP file containing school dropout data, extracts its contents, and removes the ZIP file after extraction.
}
\examples{

download_abandono_escolar()

}
