% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_CAPAG.R
\name{download_CAPAG}
\alias{download_CAPAG}
\title{Download CAPAG (Municipal Government Financial Management) Data}
\usage{
download_CAPAG(dir_path = "data_raw/")
}
\arguments{
\item{dir_path}{Character. Directory path where the downloaded file will be saved. Default is "data_raw/".
The directory will be created if it doesn't already exist.}
}
\value{
Invisible path to the downloaded file (string). The file will be named
        "CAPAG-Oficial-Municipios-2023-02-23-corrigido.xlsx" and saved in the specified directory.
}
\description{
This function downloads the CAPAG (Government Financial Management Indicator) data for municipalities
from the Tesouro Transparente platform. The data is provided in an Excel file and will be saved to the
specified directory.
}
\examples{
\dontrun{
# Download the CAPAG data to the default "data_raw/" directory
download_CAPAG()

# Download the CAPAG data to a custom directory
download_CAPAG("my_data")
}

}
