% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_IDEB_fundamental.R
\name{download_IDEB_fundamental}
\alias{download_IDEB_fundamental}
\title{Download IDEB (Índice de Desenvolvimento da Educação Básica) Data for Fundamental Education}
\usage{
download_IDEB_fundamental(dir_path = "data/")
}
\arguments{
\item{dir_path}{Path to the directory where the files will be saved and extracted.
Defaults to `"data/"`.}
}
\value{
Invisibly returns the paths of the downloaded and extracted files.
}
\description{
This function downloads two ZIP files containing IDEB data for municipalities in Brazil
for the year 2023: one for the "Anos Iniciais" (Early Years) and another for the
"Anos Finais" (Final Years). The function downloads, extracts, and removes the ZIP files
after the extraction process is completed.
}
\details{
The function performs the following steps:
1. Downloads two ZIP files containing IDEB data (one for Early Years and one for Final Years).
2. Ensures that the specified directory exists and creates it if necessary.
3. Extracts the contents of both ZIP files using the `.extract_zip` function.
4. Removes the ZIP files after extraction using the `.remove_zip` function.

The final extracted data will be saved in the specified directory, with the ZIP files
being removed after the process.
}
\examples{
\dontrun{
# Download IDEB data for Fundamental Education to the default directory ("data/")
download_IDEB_fundamental()

# Download IDEB data to a custom directory
download_IDEB_fundamental(dir_path = "data/raw/")
}

}
