% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_IVS.R
\name{download_IVS}
\alias{download_IVS}
\title{Download and Extract the IVS Data}
\usage{
download_IVS(dir_path = "data_raw/")
}
\arguments{
\item{dir_path}{A character string specifying the directory path where the data should be downloaded and extracted.
Defaults to `"data_raw/"`.}
}
\value{
This function does not return any value. It performs the download and extraction side-effects.
}
\description{
This function downloads a ZIP file from a predefined URL containing municipal data,
extracts its contents to a specified directory, and removes the ZIP file afterward.
It ensures that the directory exists, handles errors during the download and extraction process,
and uses robust extraction methods for files with special characters.
}
\details{
The function downloads a ZIP file from the IPEA API containing the basecompletamunicipal dataset.
After downloading, it attempts to extract the contents of the ZIP file using either the `zip` package (if installed)
or the base `unzip` function. After extraction, the ZIP file is deleted.

The function handles errors gracefully, displaying messages when the download or extraction fails, and cleaning up
corrupted files if necessary.
}
\examples{
# Download and extract the IVS data to the default directory
download_IVS()

# Download and extract the IVS data to a custom directory
download_IVS(dir_path = "my_data_raw/")

}
