% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_PIB.R
\name{download_PIB}
\alias{download_PIB}
\title{Download PIB Data (2010-2021) from IBGE}
\usage{
download_PIB(dir_path = "data/")
}
\arguments{
\item{dir_path}{Character. The destination directory where the downloaded ZIP file will be saved. Default is "data/".}
}
\value{
No return value, called for side effects (file download).
}
\description{
This function downloads the PIB (Produto Interno Bruto) dataset for the years 2010-2021 from the IBGE FTP server and saves it to a specified directory.
}
\examples{
# Download the PIB dataset and save it in the default directory ("data/")
download_PIB()

# Download the PIB dataset and save it in the working directory
download_PIB(".")

# Download the PIB dataset and save it in a specific folder
download_PIB("custom_folder/")

}
