% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_taxa_de_distorcao.R
\name{download_taxa_de_distorcao}
\alias{download_taxa_de_distorcao}
\title{Download and Extract Distortion Rate Data}
\usage{
download_taxa_de_distorcao(dir_path = "data_raw/")
}
\arguments{
\item{dir_path}{Character. The directory where the file will be downloaded and extracted. Default is `"data_raw/"`.}
}
\value{
No return value. The function downloads and extracts the dataset into the specified directory.
}
\description{
This function downloads the distortion rate dataset from the INEP website, extracts it, and removes the ZIP file after extraction.
}
\examples{
\dontrun{
download_taxa_de_distorcao()
download_taxa_de_distorcao("custom_directory/")
}

}
