% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_CAPAG_endividamento_2022.R
\name{extract_CAPAG_endividamento_2022}
\alias{extract_CAPAG_endividamento_2022}
\title{Extract and Process CAPAG Debt Data for 2022}
\usage{
extract_CAPAG_endividamento_2022(
  file_path = "data_raw/CAPAG/CAPAG-Oficial-Municipios-2023-02-23-corrigido.xlsx"
)
}
\arguments{
\item{file_path}{Character. The file path of the CAPAG data Excel file.
The default is 'data_raw/CAPAG/CAPAG-Oficial-Municipios-2023-02-23-corrigido.xlsx'.}
}
\value{
A data frame containing the debt information (`endividamento`) for the target cities, including
        columns: `municipio_codigo`, `municipio_nome`, `estado_sigla`, and `endividamento`.
}
\description{
This function extracts and processes the debt information (endividamento) from the CAPAG (Government Financial
Management Indicator) data for municipalities, specifically for the year 2022. The data is filtered to include
only target cities and relevant columns.
}
\details{
This function reads the CAPAG data, loads target cities data, and filters the CAPAG data to keep only the relevant
columns and rows for the specified target cities. The final result is a data frame with the debt information for 2022.
}
\examples{
\dontrun{
# Extract and process the CAPAG debt data for 2022
endividamento_2022_data <- extract_CAPAG_endividamento_2022()
}

}
