% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_dados_indicador_protecao_social.R
\name{extract_cras_data}
\alias{extract_cras_data}
\title{Trata os dados do CRAS}
\usage{
extract_cras_data(dir = "data_raw", ano = "2022")
}
\arguments{
\item{dir}{Diretório onde os dados foram baixados}

\item{ano}{Ano de interesse dos dados a serem tratados}
}
\value{
Uma tabela com os dados do CRAS a nível municipal
}
\description{
A função trata os dados do CRAS já baixados e extraídos no mesmo diretório
indicado de download
}
\examples{
\dontrun{
  extract_cras_data(dir = "data_raw", ano = "2022")
}
}
