% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_IDEB_2021_anos_finais.R
\name{extract_IDEB_2021_anos_finais}
\alias{extract_IDEB_2021_anos_finais}
\title{Extract IDEB 2021 Data for Target Cities}
\usage{
extract_IDEB_2021_anos_finais(
  file_path =
    "data/divulgacao_anos_finais_municipios_2023/divulgacao_anos_finais_municipios_2023.xlsx"
)
}
\arguments{
\item{file_path}{A character string indicating the path to the Excel file containing the IDEB
data for 2021. The default path is 'data/divulgacao_anos_finais_municipios_2023/divulgacao_anos_finais_municipios_2023.xlsx'.}
}
\value{
A data frame containing the filtered IDEB data for the target cities, with the following columns:
\itemize{
  \item municipio_codigo: The municipality code.
  \item municipio_nome: The name of the municipality.
  \item estado_sigla: The abbreviation of the state.
  \item IDEB_2021: The IDEB value for the year 2021.
}
}
\description{
This function loads the IDEB data for the year 2021 from a given Excel file, filters it
to include only the target cities specified in the `target_cities.rda` file, and returns
the filtered data with relevant columns.
}
\examples{
# Load IDEB data for target cities
ideb_data <- extract_IDEB_2021_anos_finais()

}
