% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_IDEB_2021_anos_iniciais.R
\name{extract_IDEB_2021_anos_iniciais}
\alias{extract_IDEB_2021_anos_iniciais}
\title{Extract IDEB 2021 Data for Initial Years}
\usage{
extract_IDEB_2021_anos_iniciais(
  file_path =
    "data/divulgacao_anos_iniciais_municipios_2023/divulgacao_anos_iniciais_municipios_2023.xlsx"
)
}
\arguments{
\item{file_path}{A character string specifying the path to the Excel file containing IDEB data. Default is 'data/divulgacao_anos_iniciais_municipios_2023/divulgacao_anos_iniciais_municipios_2023.xlsx'.}
}
\value{
A data frame containing filtered IDEB data with the following columns:
  - `municipio_codigo`: Municipality code
  - `municipio_nome`: Municipality name
  - `estado_sigla`: State abbreviation
  - `IDEB_2021`: IDEB score for 2021 (observed value)
}
\description{
This function extracts and processes IDEB (Índice de Desenvolvimento da Educação Básica) data for the year 2021, focusing on public schools in target municipalities.
}
\details{
The function reads the specified Excel file, processes it to keep only relevant columns, filters data for public schools, and further narrows it down to target municipalities loaded from the package dataset.
}
\examples{


ideb_data <- extract_IDEB_2021_anos_iniciais()
head(ideb_data)

}
