% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_taxa_de_distorcao_2022.R
\name{extract_taxa_de_distorcao_2022}
\alias{extract_taxa_de_distorcao_2022}
\title{Extract and Filter 2022 Distortion Rate Data}
\usage{
extract_taxa_de_distorcao_2022(
  file_path = "data/TDI_2022_MUNICIPIOS/TDI_MUNICIPIOS_2022.xlsx"
)
}
\arguments{
\item{file_path}{Character. The path to the Excel file containing the 2022 distortion rate data. Default is `"data/TDI_2022_MUNICIPIOS/TDI_MUNICIPIOS_2022.xlsx"`.}
}
\value{
A data frame containing the filtered distortion rate data for target cities.
}
\description{
This function loads the 2022 distortion rate dataset from an Excel file, filters the relevant data, and returns only the records for target cities.
}
\examples{
\dontrun{
extract_taxa_de_distorcao_2022()
extract_taxa_de_distorcao_2022("custom_path/TDI_2022.xlsx")
}

}
