% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_taxa_matricula_creche.R
\name{extract_taxa_matricula_creche}
\alias{extract_taxa_matricula_creche}
\title{Extract Early Childhood Education Enrollment Rate Data for Target Cities}
\usage{
extract_taxa_matricula_creche(
  file_path = "data/Taxa líquida de matrículas em creches.xlsx"
)
}
\arguments{
\item{file_path}{Character. Path to the Excel file containing the enrollment data. Default is
'data/Taxa líquida de matrículas em creches.xlsx'. The function will load the data from the file
and filter for the year 2022, focusing on target cities.}
}
\value{
A data frame with the following columns:
        - `municipio_codigo`: The IBGE code of the municipality.
        - `municipio_nome`: The name of the municipality.
        - `taxa_de_matricula_2022`: The enrollment rate for the year 2022 in the target cities.
}
\description{
This function loads and processes data from an Excel file containing early childhood education
enrollment rates (taxa de matrícula em creches) for various municipalities, specifically for the year 2022.
The data is filtered to include only municipalities listed in the target cities data.
}
\examples{
\dontrun{
# Load and process data from the default file
taxa_matricula_data <- extract_taxa_matricula_creche()

# Load and process data from a custom file path
taxa_matricula_data <- extract_taxa_matricula_creche("custom_data/Taxa líquida de matrículas em creches.xlsx")
}

}
