% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abastecimento_esgoto.R
\name{get_abastecimento_esgoto}
\alias{get_abastecimento_esgoto}
\title{Extração dos dados de Abastecimento de Água e Esgotamento Sanitário}
\usage{
get_abastecimento_esgoto()
}
\value{
Uma dataframe
}
\description{
A função extrai os dados de Abastecimento de Água e Esgotamento Sanitário a
nível municipal para o ano de 2021 diretamente da plataforma do Instituto Água
e Saneamento.
}
\examples{
\dontrun{
  data_abatecimento_e_esgoto <- get_abastecimento_esgoto()
}
}
