% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alimentos_pnae.R
\name{get_alimentos_pnae}
\alias{get_alimentos_pnae}
\title{Extração dos dados de Aquisição Agricultura Familiar pelo PNAE}
\usage{
get_alimentos_pnae(dir = "data_raw/pnae", ano = 2022)
}
\arguments{
\item{dir}{Diretório onde os arquivos serão salvos ou lidos.}

\item{ano}{Ano de interesse dos dados}
}
\value{
Um dataframe.
}
\description{
A função extrai e trata os dados de Aquisição Agricultura Familiar pelo PNAE de um
determinado ano a nível municipal. Os dados são extraídos diretamente do Fundo
Nacional de Desenvolvimento da Educação do gov.br.
}
\examples{
\dontrun{
  data_prop_alimentos_pnae <- get_alimentos_pnae(
    dir = "data/pnae",
    ano = 2022
  )
}
}
