% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cad_bf_data.R
\name{get_cad_bf_data}
\alias{get_cad_bf_data}
\title{Extração dos dados de cadastros do CadÚnico e Bolsa Família}
\usage{
get_cad_bf_data(
  cod_municipios_ibge = target_cities$municipio_codigo,
  ano = "2023",
  mes = "06"
)
}
\arguments{
\item{cod_municipios_ibge}{Lista com o código dos munícipios do IBGE. Como padrão
utiliza a coluna municipios codigo da tabela target_cities.}

\item{ano}{Ano de interesse. Por padrão usa o ano de 2023}

\item{mes}{Mês de interesse. Por padrão usa o mes de junho (06)}
}
\value{
Um data.frame
}
\description{
Essa função extrai os dados de unidades de famílias cadastradas no CadÚnico e
Bolsa Família a nível municipal segundo a plataforma do Aplicações Cidadania.
}
\examples{
\dontrun{
  get_cad_bf_data(
    cod_municipios_ibge = c("1508126", "1708205", "2101400", "2112001"),
    ano = "2020",
    mes = "07"
   )
}
}
