% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cobertura_ab.R
\name{get_cobertura_ab}
\alias{get_cobertura_ab}
\title{Extração dos dados de Cobertura de Atenção Básica}
\usage{
get_cobertura_ab(ano = 2020)
}
\arguments{
\item{ano}{Ano de interesse dos dados. Por padrão usa o ano de 2020}
}
\value{
Um dataframe
}
\description{
A função extrai os dados de Cobertura de Atenção Básica diretamente da plataforma
e-Gestor do ministério da Saúde. Os dados são detalhados a nível municipal para
o ano indicado.
}
\examples{
\dontrun{
  dados_cobertura_ab <- get_cobertura_ab(
    ano = "2020"
  )
}
}
