% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cobertura_vacinal.R
\name{get_cobertura_vacinal}
\alias{get_cobertura_vacinal}
\title{Extrai os dados de Cobertura Vacinal}
\usage{
get_cobertura_vacinal(cod_ibge = target_cities$municipio_codigo)
}
\arguments{
\item{cod_ibge}{Vetor character com os códigos do ibge dos municípios. Por padrão
utiliza os códigos dos municípios da tabela target_cities}
}
\value{
Um dataframe
}
\description{
A função extrai os dados de Cobertura Vacinal a nível municipal diretamente
da plataforma da FGV Municípios
}
\examples{
\dontrun{
 data_cobertura_vacinal <- get_cobertura_vacinal(
  cod_ibge = c("1508126", "1708205", "2101400")
 )
}
}
