% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conectividade_rural.R
\name{get_conectividade_rural}
\alias{get_conectividade_rural}
\title{Pega Indicador de Conectividade Rural (ICR) por Estado}
\usage{
get_conectividade_rural(sigla_uf = NULL)
}
\arguments{
\item{sigla_uf}{Character vector. Sigla(s) dos estados (UFs) a serem consultados.
Se NULL (default), busca dados para todos os estados ("AC", "AL", ..., "TO").}
}
\value{
Um tibble contendo os dados de conectividade rural para os municípios do(s) estado(s) informado(s).
As colunas retornadas dependem da estrutura do JSON fornecido pelo site.
}
\description{
Esta função acessa o site da ConectarAgro e baixa os indicadores de conectividade rural
para o(s) estado(s) selecionado(s). Se nenhum estado for especificado, a função retorna dados para todos os estados brasileiros.
}
\examples{
\dontrun{
# Buscar dados apenas do Acre
pega_conectividade_rural("AC")

# Buscar dados de todos os estados
pega_conectividade_rural()
}

}
