% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deslocamento_SIDRA.R
\name{get_deslocamento_SIDRA}
\alias{get_deslocamento_SIDRA}
\title{Extração dos dados de Tempo de Deslocamento de Casa/Trabalho}
\usage{
get_deslocamento_SIDRA(ano = 2010)
}
\arguments{
\item{ano}{Ano de interesse dos dados. Por padrão utiliza 2010 devido a ser o
único ano disponível}
}
\value{
Uma dataframe
}
\description{
Essa função extrai os dados de Tempo Habitual de deslocamento de Casa para o
Trabalho a nível municípal de um determinado ano diretamente da API do SIDRA.
}
\examples{
\dontrun{
 dados_deslocamento <- get_deslocamento_SIDRA(ano = 2010)
}
}
