% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_emissao.R
\name{get_emissao}
\alias{get_emissao}
\title{Extrai os dados de Emissão por População}
\usage{
get_emissao(ano = 2019)
}
\arguments{
\item{ano}{Ano de interesse dos dados. Por padrão usa o mais recente, 2019.}
}
\value{
Retorna uma tabela com os dados de Emissão por Populção a nível municipal
}
\description{
A função extra os dados de Emissão por População do SEEG Municípios.
}
\examples{
\dontrun{
  data <- get_emissao()
}
}
