% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_empregos_formais_por_100khab.R
\name{get_empregos_formais_por_100khab}
\alias{get_empregos_formais_por_100khab}
\title{Obter Empregos Formais por 100k Habitantes}
\usage{
get_empregos_formais_por_100khab()
}
\value{
Um data frame contendo os dados de empregos formais por 100.000 habitantes para os municípios.
        As colunas retornadas incluem:
        \itemize{
          \item `municipio_codigo`: Código único do município.
          \item `municipio_nome`: Nome do município.
          \item `estado_sigla`: Sigla do estado.
          \item `populacao`: População do município.
          \item `empregos_por_100khab`: Empregos formais por 100.000 habitantes no município.
        }
}
\description{
Esta função retorna a quantidade de empregos formais por 100.000 habitantes para os municípios,
com base nos dados de empregos formais de 2022 e a população dos municípios.
}
\details{
Ela realiza o seguinte processo:
\itemize{
  \item Carrega os dados de empregos formais de 2022 usando a função `get_empregos_formais_total_2022()`.
  \item Carrega os dados de população dos municípios e realiza as transformações necessárias, como a criação da coluna `municipio_codigo`, remoção das colunas `cod_uf` e `cod_munic`, e renomeação de colunas.
  \item Realiza o `left_join` entre os dados de empregos formais e a população, e calcula a quantidade de empregos formais por 100.000 habitantes.
}
}
\examples{
# Exemplo de uso:
empregos_por_100khab <- get_empregos_formais_por_100khab()
head(empregos_por_100khab)

}
