% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_empregos_formais_total_2022.R
\name{get_empregos_formais_total_2022}
\alias{get_empregos_formais_total_2022}
\title{Obter Dados de Empregos Formais Totais para 2022}
\usage{
get_empregos_formais_total_2022()
}
\value{
Um data frame contendo o código do município (`municipio_codigo`), nome do município (`municipio_nome`),
        sigla do estado (`estado_sigla`) e o total de empregos formais (`total_empregos`) para as cidades-alvo.
}
\description{
Esta função processa os dados da RAIS (Relação Anual de Informações Sociais) para retornar
o número total de empregos formais para cidades-alvo específicas em 2022.
Ela faz o download e lê os dados da RAIS, filtra-os com base nas cidades-alvo predefinidas,
e retorna as informações relevantes.
}
\examples{
# Obter total de empregos formais para as cidades-alvo em 2022
total_empregos_2022 <- get_empregos_formais_total_2022()
head(total_empregos_2022)
}
