% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_endividamento.R
\name{get_endividamento}
\alias{get_endividamento}
\title{Obtém dados municipais de Dívida Consolidada/Receita Corrente Líquida (Endividamento)}
\usage{
get_endividamento(download = TRUE, ano = 2025)
}
\arguments{
\item{download}{Lógico. Se `TRUE`, baixa o arquivo correspondente ao ano escolhido. Default é `TRUE`.}

\item{ano}{Inteiro ou caractere. Ano desejado para a obtenção dos dados. Valores válidos: 2017 a 2023. Default é 2023.}
}
\value{
Um `data.frame` contendo os dados de endividamento dos municípios para o ano especificado, incluindo os códigos dos municípios, nomes e estados.
}
\description{
Esta função baixa e processa os dados de endividamento municipal a partir dos arquivos CAPAG.
}
\examples{

# Obtendo dados para 2023 (sem baixar novamente se o arquivo já existir)
df <- get_endividamento(download = FALSE, ano = 2023)

# Obtendo dados para 2021, forçando o download
df <- get_endividamento(download = TRUE, ano = 2021)

}
