% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_equip_estab.R
\name{get_equip_estab}
\alias{get_equip_estab}
\title{Extrai os dados de Equipamento por Estabelecimento do Tabnet DATASUS}
\usage{
get_equip_estab(ano = "25", mes = "02")
}
\arguments{
\item{ano}{Ano de interesse dos dados. Por padrão usa os dados mais recentes.}

\item{mes}{Mes de interesse dos dados. Por padrão usa os dados mais recentes.}
}
\value{
Um dataframe
}
\description{
A função extrai os dados de Equipamentos por Estabelecimento e Número de Equipamentos
Existentes de cada município para um determinado ano e mês. Os dados são extraídos
diretamente do CNES - EQUIPOBR da plataforma Tabnet DATASUS
}
\examples{
\dontrun{
  equip_por_estab_mun <- get_equip_estab(ano = "23", mes = "04")
}
}
