% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estab_dirigido_por_mulheres.R
\name{get_estab_dirigido_por_mulheres}
\alias{get_estab_dirigido_por_mulheres}
\title{Extrai os dados de Estabelecimentos Agropecuários dirigidos por Mulheres}
\usage{
get_estab_dirigido_por_mulheres(cod_ibge = "Brasil")
}
\arguments{
\item{cod_ibge}{Um vetor de caracteres com o código IBGE do município ou "Brasil" para todos os municípios ou "Targets" para os municípios alvo.}
}
\value{
Um dataframe
}
\description{
A função extrai  dados de Estabelecimentos Agropecuários dirigidos por Mulheres
do Senso Agro de 2017 através da API do SIDRA IBGE.
}
\examples{
\dontrun{
  data <- get_dirigido_por_mulheres()
}
}
