% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_familias_agricultura_familiar_cadastradas_cadunico.R
\name{get_familias_agricultura_familiar_cadastradas_cadunico}
\alias{get_familias_agricultura_familiar_cadastradas_cadunico}
\title{Extrai os dados de Famílias de Agricultores Familiar que são Cadastrados no CadÚnico}
\usage{
get_familias_agricultura_familiar_cadastradas_cadunico(
  cod_ibge = target_cities$municipio_codigo,
  ano = "last",
  mes = "01"
)
}
\arguments{
\item{cod_ibge}{Vetor character com os códigos do ibge dos municípios. Por padrão
utiliza os códigos dos municípios da tabela target_cities}

\item{ano}{Ano de interesse dos dados. Por padrão usa "last" que se refere ao
ano dado mais recente.}

\item{mes}{Mês de interesse dos dados. O mês deve ser um character entre "01"
e "12"}
}
\value{
Um dataframe
}
\description{
Essa função extrai os dados de Famílias de Agricultores Familiar que são Cadastrados
no CadÚnico diretamente da plataforma do Cidadania a nível municipal para um
determinado ano e mês.
}
\examples{
\dontrun{
  data <- get_familias_agricultura_familiar_cadastradas_cadunico(
    cod_ibge = c("1508126", "1708205", "2101400"),
    ano = 2024,
    mes = "03"
  )
}
}
