% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_dados_indicador_educacao.R
\name{get_gastos_per_capita_educacao}
\alias{get_gastos_per_capita_educacao}
\title{Obter os gastos per capita em educação por município}
\usage{
get_gastos_per_capita_educacao(cod_ibge = target_cities$municipio_codigo)
}
\arguments{
\item{cod_ibge}{Vetor de códigos IBGE dos municípios a serem consultados.
O padrão é `target_cities$municipio_codigo`. Deve ser um vetor numérico.}
}
\value{
Um `data.frame` contendo os dados de gastos per capita em educação
        para os municípios consultados, incluindo as colunas:
        - `municipio_codigo`: Código IBGE do município.
        - `ano`: Ano da observação.
        - `gastos_per_capita_educacao`: Valor do gasto per capita em educação.
        - Demais colunas retornadas pela API, com nomes padronizados.
}
\description{
Esta função coleta os dados de gastos per capita em educação para os municípios
especificados, consultando a API da FGV.
}
\examples{

# Obter os gastos per capita em educação para os municípios padrão
get_gastos_per_capita_educacao()

# Obter os gastos per capita em educação para municípios específicos
get_gastos_per_capita_educacao(c(3550308, 3304557))

}
