% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_dados_indicador_saude.R
\name{get_gastos_per_capta_saude}
\alias{get_gastos_per_capta_saude}
\title{Extração dos dados de Gasto Per Capta com Saúde}
\usage{
get_gastos_per_capta_saude(
  cod_municipios_ibge = target_cities$municipio_codigo
)
}
\arguments{
\item{cod_municipios_ibge}{Codigos do ibge para os muncípios}
}
\value{
Uma tabela com os dados de Gasto Per Capta com Saúde para os anos disponíveis
}
\description{
A função extra is dados de Gasto Per Capta com Saúde para os municípios direto
da plataforma da FGV Municípios
}
\examples{
\dontrun{
  get_gastos_per_capta_saude(cod_municipios_ibge = c("1508126", "1708205", "2101400"))
}
}
