% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_dados_indicador_educacao.R
\name{get_gastos_totais_educacao}
\alias{get_gastos_totais_educacao}
\title{Obter gastos totais com educação por município}
\usage{
get_gastos_totais_educacao(ano = 2022)
}
\arguments{
\item{ano}{Integer. Ano dos dados a serem carregados (entre 2013 e 2024). Default: 2022.}
}
\value{
Dataframe com os municípios e seus respectivos gastos totais com educação.
}
\description{
Esta função carrega e processa os dados de gastos públicos com educação
a partir de arquivos CSV, filtrando por município e consolidando os valores.
}
\details{
Os dados foram retirados de https://siconfi.tesouro.gov.br/siconfi/pages/public/consulta_finbra/finbra_list.jsf
}
\examples{
gastos_educacao <- get_gastos_totais_educacao(ano = 2023)
head(gastos_educacao)

}
