% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gestao_dr_vs_de.R
\name{get_gestao_dr_vs_de}
\alias{get_gestao_dr_vs_de}
\title{Obtém o indicador Gestão (Dispêndio do Recurso vs Desempenho - Educação)}
\usage{
get_gestao_dr_vs_de(ano = 2023, download_perform = FALSE)
}
\arguments{
\item{ano}{Ano de referência (padrão: 2023).}

\item{download_perform}{Se TRUE baixa os dados do ideb. Default é FALSE.}
}
\value{
Um data frame contendo o código do município, nome, estado e o índice de gestão de recursos vs desempenho educacional.
}
\description{
Obtém o indicador Gestão (Dispêndio do Recurso vs Desempenho - Educação)
}
\examples{
\dontrun{
  get_gestao_dr_vs_de(2023)
}

}
