% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IDEB_2021_anos_finais.R
\name{get_IDEB_2021_anos_finais}
\alias{get_IDEB_2021_anos_finais}
\title{Obtém os dados do IDEB 2021 para os anos finais do ensino fundamental}
\usage{
get_IDEB_2021_anos_finais(
  download = TRUE,
  file_path =
    "data_raw/divulgacao_anos_finais_municipios_2023/divulgacao_anos_finais_municipios_2023.xlsx"
)
}
\arguments{
\item{download}{Lógico. Se `TRUE`, a função baixa o conjunto de dados do IDEB antes do processamento.
O padrão é `TRUE`.}

\item{file_path}{Caracter. Caminho para o arquivo Excel contendo os dados do IDEB.
O padrão é `'data_raw/divulgacao_anos_finais_municipios_2023/divulgacao_anos_finais_municipios_2023.xlsx'`.}
}
\value{
Um data frame contendo os dados processados do IDEB para os anos finais do ensino fundamental.
}
\description{
Esta função obtém os dados do IDEB (Índice de Desenvolvimento da Educação Básica)
para os anos finais do ensino fundamental em municípios brasileiros no ano de 2021.
Caso necessário, a função realiza o download do conjunto de dados antes do processamento.
}
\details{
A função segue os seguintes passos:
1. Se `download = TRUE`, chama `download_IDEB_fundamental()` para baixar e extrair os dados do IDEB.
2. Chama `extract_IDEB_2021_anos_finais()` para processar e retornar os dados filtrados.
}
\examples{
\dontrun{
# Baixar e obter os dados do IDEB para os anos finais
ideb_dados <- get_IDEB_2021_anos_finais()

# Utilizar um arquivo já existente sem baixar novamente
ideb_dados <- get_IDEB_2021_anos_finais(download = FALSE, file_path = "caminho_personalizado.xlsx")
}

}
