% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IDEB_fundamental.R
\name{get_IDEB_fundamental}
\alias{get_IDEB_fundamental}
\title{Obter Dados do IDEB para o Ensino Fundamental}
\usage{
get_IDEB_fundamental(download = TRUE, ano = "2023", tipo = "anos_finais")
}
\arguments{
\item{download}{Um valor lógico (TRUE ou FALSE). Se TRUE, a função fará o download dos arquivos caso eles não existam
localmente. Caso contrário, apenas tentará carregar os dados já existentes.}

\item{ano}{Um valor numérico ou caractere indicando o ano para o qual os dados devem ser obtidos. Os anos disponíveis
são: 2005, 2007, 2009, 2011, 2013, 2015, 2017, 2019, 2021 e 2023.}

\item{tipo}{Um valor caractere, podendo ser "anos_finais" ou "anos_iniciais". Define se os dados são dos anos finais
ou iniciais do Ensino Fundamental.}
}
\value{
Um data frame contendo os dados do IDEB filtrados, com as colunas: `municipio_codigo`, `municipio_nome`,
  `estado_sigla` e o valor do IDEB para o ano solicitado.
}
\description{
Esta função permite obter os dados do IDEB (Índice de Desenvolvimento da Educação Básica) para o Ensino Fundamental,
tanto para os anos finais quanto para os anos iniciais. Os dados podem ser baixados automaticamente, caso não estejam
presentes no diretório local.
}
\examples{
\dontrun{
  # Obter dados do IDEB para os anos finais de 2023
  ideb_finais <- get_IDEB_fundamental(download = TRUE, ano = "2023", tipo = "anos_finais")

  # Obter dados do IDEB para os anos iniciais de 2023
  ideb_iniciais <- get_IDEB_fundamental(download = TRUE, ano = "2023", tipo = "anos_iniciais")
}

}
