% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_internet_data.R
\name{get_internet_data}
\alias{get_internet_data}
\title{Pega os dados de Telecomunicações dos Municipios Brasileiros}
\usage{
get_internet_data(dir = "data_raw", ano = 2024)
}
\arguments{
\item{dir}{Diretório onde os dados serão baixados. Por padrão utiliza "data_raw"}

\item{ano}{Ano de interesse dos dados. Por padrão utliza 2024.}
}
\value{
Um dataframe
}
\description{
Essa função pega os dados de Telecomunicações dos Municipios Brasileiros direto
dos dados abertos da ANATEL, salva os dados baixados no diretório destinado na
pasta "internet_data_anatel" e retorna um dataframe com os dados por município
no ano indicado.
}
\examples{
\dontrun{
  data <- get_internet_data(dir = "data", ano = 2022)
}
}
