% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_investimento_per_capita.R
\name{get_investimento_per_capita}
\alias{get_investimento_per_capita}
\title{Extrai os dados da nota quanto a Investimento Per Capita do CFA}
\usage{
get_investimento_per_capita(
  cod_ibge = target_cities$municipio_codigo,
  ano = 2024
)
}
\arguments{
\item{cod_ibge}{Vetor character com os códigos do ibge dos municípios. Por padrão
utiliza os códigos dos municípios da tabela target_cities}

\item{ano}{Ano de interesse dos dados. Por padrão usa o ano de 2024.}
}
\value{
Um dataframe
}
\description{
A função extrai os dados da nota quanto a Investimento Per Capita do CFA a
nível municipal diretamente do PowerBI Índice de Governança Municipal do Conselho
Federal de Administração, IGM - CFA.
}
\examples{
\dontrun{
 data_investimento_per_capita <- get_cobertura_vacinal(
  cod_ibge = c("1508126", "1708205", "2101400"),
  ano = 2024
 )
}
}
