% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_liquidez.R
\name{get_liquidez}
\alias{get_liquidez}
\title{Obtém dados municipais de Obrigações Financeiras/Disponibilidade de Caixa (Liquidez)}
\usage{
get_liquidez(download = TRUE, ano = 2023)
}
\arguments{
\item{download}{Lógico. Se `TRUE`, baixa o arquivo correspondente ao ano escolhido. Default é `TRUE`.}

\item{ano}{Inteiro ou caractere. Ano desejado para a obtenção dos dados. Valores válidos: 2017 a 2023. Default é 2023.}
}
\value{
Um `data.frame` contendo os dados de liquidez dos municípios para o ano especificado, incluindo os códigos dos municípios, nomes e estados.
}
\description{
Esta função baixa e processa os dados de liquidez municipal a partir dos arquivos CAPAG.
}
\examples{

# Obtendo dados para 2023 (sem baixar novamente se o arquivo já existir)
df <- get_liquidez(download = FALSE, ano = 2023)

# Obtendo dados para 2021 e assegurando a realização do download
df <- get_liquidez(download = TRUE, ano = 2021)

}
