% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obitos_DATASUS.R
\name{get_obitos_DATASUS}
\alias{get_obitos_DATASUS}
\title{Extração dos dados de Óbitos por acidentes de trânsito por munícipio do DATASUS}
\usage{
get_obitos_DATASUS(ano)
}
\arguments{
\item{ano}{Ano dos dados que serão requisitados. É necessário colocar apenas a década
do ano de interesse no formato string, por exemplo: 2008, ano = "08"}
}
\value{
Uma tabela com nome do município, quantidade de óbitos, código
do estado segundo IBGE e o ano dos dados
}
\description{
A função extrai os dados de óbitos por acidente de trânsito a nível municipal
por ano da plataforma do DATASUS. A função retorna uma tabela com nome do município,
quantidade de óbitos, código do estado segundo IBGE e o ano dos dados
}
\examples{
\dontrun{
  get_obitos_DATASUS(ano = "09")
}
}
