% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obitos_evitaveis.R
\name{get_obitos_evitaveis}
\alias{get_obitos_evitaveis}
\title{Extração dados de Óbitos por Causas Evitáveis DATASUS}
\usage{
get_obitos_evitaveis(ano = 2023)
}
\arguments{
\item{ano}{Ano de interesse dos dados}
}
\value{
Um dataframe com os dados óbitos evitáveis e as causas evitáveis a nível municipal
}
\description{
A função extrai os dados óbitos por Causas evitáveis em menores do que 5 anos e
óbitos por Causas evitáveis entre 5 a 75 anos direto da plataforma do DATASUS
através do pacote datasus a nível municipal.
}
\examples{
\dontrun{
  get_obitos_evitaveis(ano = "2023")
}
}
