% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paridades_camara_mun.R
\name{get_paridades_camara_mun}
\alias{get_paridades_camara_mun}
\title{Pega os dados de Paridade de Gênero e Paridade de Negros das Câmaras Municipais -
Indicador de Proteção Social}
\usage{
get_paridades_camara_mun()
}
\value{
Um dataframe
}
\description{
Essa função pega os dados de Paridade de Gênero e Paridade de Negros de Câmaras
Municipais diretamente do TSE
}
\examples{
\dontrun{
  data <- get_paridades_camara_mun()
}
}
