% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_perc_agricultor_ens_bas.R
\name{get_perc_agricultor_ens_bas}
\alias{get_perc_agricultor_ens_bas}
\title{Extrai os dados de Estabelecimentos Agropecuários em que o Agricultor concluiu o Ensino Básico}
\usage{
get_perc_agricultor_ens_bas()
}
\value{
Um dataframe
}
\description{
A função extrai  dados de Estabelecimentos Agropecuários em que o Agricultor
concluiu o Ensino Básico do Senso Agro de 2017 através da API do SIDRA IBGE.
}
\examples{
\dontrun{
  data <- get_perc_agricultor_ens_bas()
}
}
