% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_perc_area_cultv_espec_flor.R
\name{get_perc_area_cultv_espec_flor}
\alias{get_perc_area_cultv_espec_flor}
\title{Extrai os dados da Área Cultivada com Espécies Florestais}
\usage{
get_perc_area_cultv_espec_flor(cod_ibge = "Brasil")
}
\arguments{
\item{cod_ibge}{Um vetor de caracteres com o código IBGE do município ou "Brasil" para todos os municípios ou "Targets" para os municípios alvo.}
}
\value{
Um dataframe
}
\description{
A função extrai  dados da Área Cultivada com Espécies Florestais do Senso Agro
de 2017 através da API do SIDRA IBGE.
}
\examples{
\dontrun{
  data <- get_perc_area_cultv_espec_flor()
}
}
