% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_perc_empresas_permanecem_abertas.R
\name{get_perc_empresas_permanecem_abertas}
\alias{get_perc_empresas_permanecem_abertas}
\title{Tratamento dos dados de Empresas que permanecem permanecem abertas}
\usage{
get_perc_empresas_permanecem_abertas()
}
\value{
Um dataframe.
}
\description{
A função trata os dados de Empresas que permanecem abertas que foram extraídos
diretamente do Painel do Mapa de Empresas do gov.br
}
\examples{
\dontrun{
 data_empresas <- get_perc_empresas_permanecem_abertas()
}
}
