% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_perc_prod_jovens_25_35_anos.R
\name{get_perc_prod_jovens_25_35_anos}
\alias{get_perc_prod_jovens_25_35_anos}
\title{Extrai os dados de Estabelecimento Agropecuário dirigidos por Jovens}
\usage{
get_perc_prod_jovens_25_35_anos()
}
\value{
Um dataframe
}
\description{
A função extrai  dados de Estabelecimento Agropecuário dirigidos por Jovens
entre 25 e 35 anos do Senso Agro de 2017 através da API do SIDRA IBGE.
}
\examples{
\dontrun{
  data <- get_perc_prod_jovens_25_35_anos()
}
}
