% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_percentual_acess_agua_mun.R
\name{get_percentual_acess_agua_mun}
\alias{get_percentual_acess_agua_mun}
\title{Obtém o percentual de domicílios com acesso à rede geral de água}
\usage{
get_percentual_acess_agua_mun(cod_ibge = "Brasil")
}
\arguments{
\item{cod_ibge}{Um vetor de caracteres com o código IBGE do município ou "Brasil" para todos os municípios ou "Targets" para os municípios alvo.}
}
\value{
Um tibble com o percentual de domicílios com acesso à rede geral de água por município e ano.
}
\description{
A função consulta a API do SIDRA/IBGE para obter dados sobre abastecimento de água e calcula
o percentual de domicílios com ligação à rede geral como forma principal de abastecimento.
}
\examples{
\dontrun{
 percentual_acess_agua_mun_data <- get_percentual_acess_agua_mun()
}

}
