% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_percentual_estabelecimentos_agricultura_familiar.R
\name{get_percentual_estabelecimentos_agricultura_familiar}
\alias{get_percentual_estabelecimentos_agricultura_familiar}
\title{Proporção de estabelecimentos de agricultura familiar - Desenvolvimento Econômico}
\usage{
get_percentual_estabelecimentos_agricultura_familiar(cod_ibge = "Brasil")
}
\arguments{
\item{cod_ibge}{Um vetor de caracteres com o código IBGE do município ou "Brasil" para todos os municípios ou "Targets" para os municípios alvo.}
}
\value{
Um dataframe
}
\description{
Essa função pega os dados da Proporção de Estabelecimentos de Agricultura Familiar
diretamente do Censo Agro na plataforma do SIDRA
}
\examples{
\dontrun{
  data <- get_percentual_estabelecimentos_agricultura_familiar()
}
}
