% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_percentual_vias_pav_mun.R
\name{get_percentual_vias_pav_mun}
\alias{get_percentual_vias_pav_mun}
\title{Obtém o percentual de vias pavimentadas nos municípios selecionados}
\usage{
get_percentual_vias_pav_mun(cod_ibge = "Brasil")
}
\arguments{
\item{cod_ibge}{Um vetor de caracteres com o código IBGE do município ou "Brasil" para todos os municípios ou "Targets" para os municípios alvo.}
}
\value{
Um tibble com o percentual de domicílios em vias pavimentadas por município e ano.
}
\description{
Esta função coleta dados da Pesquisa Urbanística do Entorno dos Domicílios via API do SIDRA/IBGE
e calcula o percentual de domicílios localizados em setores com vias pavimentadas.
}
\examples{
\dontrun{
 percentual_vias_pav_mun_data <- get_percentual_vias_pav_mun()
}
}
