% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PIB_per_capta.R
\name{get_PIB_per_capta}
\alias{get_PIB_per_capta}
\title{Obter PIB per capita dos municípios}
\usage{
get_PIB_per_capta(download = TRUE, ano = 2021)
}
\arguments{
\item{download}{Lógico. Se `TRUE`, a função irá tentar baixar os dados mais recentes
caso o arquivo não esteja presente no diretório local. O padrão é `TRUE`.}

\item{ano}{Ano para o qual o PIB per capita será obtido. O valor padrão é `2021`.
Os anos válidos são de 2010 a 2021.}
}
\value{
Um data frame contendo o código do município, a sigla da unidade da federação,
o nome do município e o PIB per capita para o ano especificado.
}
\description{
Esta função permite obter o PIB per capita dos municípios de um ano específico.
Se o arquivo de dados não estiver disponível, ele será baixado automaticamente.
}
\examples{
\dontrun{
  pib_data <- get_PIB_per_capta(ano = 2021)
  head(pib_data)
}

}
