% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop_ocup_mun.R
\name{get_pop_ocup_mun}
\alias{get_pop_ocup_mun}
\title{População ocupada}
\usage{
get_pop_ocup_mun(cod_ibge)
}
\arguments{
\item{cod_ibge}{Vetor com os códigos dos municípios do IBGE.}
}
\value{
Um dataframe
}
\description{
Essa função extrai os dados de População Ocupada diretamente da API do IBGE.
}
