% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_poupanca_corrente.R
\name{get_poupanca_corrente}
\alias{get_poupanca_corrente}
\title{Obtém dados da Despesa Corrente/Receita Corrente Ajustada (Poupança Corrente) para um ano específico}
\usage{
get_poupanca_corrente(download = TRUE, ano = 2023)
}
\arguments{
\item{download}{Lógico. Se \code{TRUE}, a função tentará fazer o download dos dados se o arquivo
não estiver presente localmente. O padrão é \code{TRUE}.}

\item{ano}{Um número ou caractere. O ano para o qual os dados devem ser obtidos. O valor padrão é
2023, mas os anos válidos são de 2017 a 2023.}
}
\value{
Um \code{data.frame} contendo os dados da poupança corrente dos municípios alvo,
   com as colunas:
   \itemize{
     \item \code{municipio_codigo}: Código do município.
     \item \code{municipio_nome}: Nome do município.
     \item \code{estado_sigla}: Sigla do estado do município.
     \item \code{poupanca_corrente_<ano>}: O valor da poupança corrente para o ano especificado.
   }
}
\description{
Esta função carrega e processa os dados da poupança corrente para os municípios brasileiros,
a partir de arquivos CAPAG disponíveis para os anos de 2017 a 2023. A função permite o
download automático dos dados caso o arquivo não esteja presente no diretório especificado.
}
\details{
O arquivo de dados CAPAG utilizado para cada ano é obtido automaticamente a partir de um
   diretório local. Caso o arquivo não exista, a função faz o download do mesmo. Para o ano de 2023,
   um ajuste no cabeçalho da planilha é feito (saltando 2 linhas de dados).
}
\examples{
\dontrun{
# Obter dados para o ano de 2021
poupanca_corrente_2021 <- get_poupanca_corrente(ano = 2021)

# Obter dados para o ano de 2023 e sem download
poupanca_corrente_2023 <- get_poupanca_corrente(download = FALSE, ano = 2023)
}

}
