% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sal_med_mun.R
\name{get_sal_med_mun}
\alias{get_sal_med_mun}
\title{Salário médio mensal dos trabalhadores formais}
\usage{
get_sal_med_mun(cod_ibge)
}
\arguments{
\item{cod_ibge}{Vetor com os códigos dos municípios do IBGE.}
}
\value{
Um dataframe
}
\description{
A função extrai o Salário médio mensal dos trabalhadores formais diretamente
da API do IBGE.
}
