% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_dados_indicador_protecao_social.R
\name{get_situacao_pobreza_data}
\alias{get_situacao_pobreza_data}
\title{Extração o número de familias e pessoas em situação de pobreza do Cadúnico}
\usage{
get_situacao_pobreza_data(cod_municipios_ibge = target_cities$municipio_codigo)
}
\arguments{
\item{cod_municipios_ibge}{Lista com o código dos munícipios do IBGE. Como padrão
utiliza a coluna municipios codigo da tabela target_cities.}
}
\value{
Um data.frame
}
\description{
Essa função extrai o número de pessoas e famílias em situação de pobreza e que
estão cadastradas no CadÚnico da plataforma Aplicações Cidadania.
}
\examples{
\dontrun{
 get_situacao_pobreza_data(cod_municipios_ibge = c("1508126", "1708205", "2101400"))
}
}
