% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_dados_indicador_educacao.R
\name{get_taxa_liq_matricula_creche}
\alias{get_taxa_liq_matricula_creche}
\title{Obter Taxa Líquida de Matrícula em Creches por Município}
\usage{
get_taxa_liq_matricula_creche(ano = "2023")
}
\arguments{
\item{ano}{Um valor numérico ou caractere indicando o ano para o qual os dados devem ser obtidos. Os anos disponíveis
são de 2010 até 2023. O padrão é "2023".}
}
\value{
Um data frame contendo os dados da taxa líquida de matrícula em creches, com as colunas: `municipio_codigo`,
  `municipio_nome`, `taxa_liquida_matricula_creche_<ano>`, onde `<ano>` é o ano solicitado.
}
\description{
Esta função permite obter os dados da taxa líquida de matrícula em creches fornecidos pelo INEP para o ano especificado.
Os dados são carregados a partir de um arquivo local.
}
\examples{
\dontrun{
  # Obter taxa líquida de matrícula em creches para o ano de 2023
  taxa_2023 <- get_taxa_matricula_creche("2023")

  # Obter taxa líquida de matrícula em creches para o ano de 2022
  taxa_2022 <- get_taxa_matricula_creche("2022")
}

}
