% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_dados_indicador_saude.R
\name{get_taxa_mortalidade_infantil}
\alias{get_taxa_mortalidade_infantil}
\title{Extração dados de Mortalidade Infantil e Nascimento DATASUS}
\usage{
get_taxa_mortalidade_infantil(ano = "2022")
}
\arguments{
\item{ano}{Ano de interesse dos dados}
}
\value{
Um dataframe com os dados de Mortalidade Infantil, Nascimento, Taxa
de Mortaliade Infantil para 1000 nascimentos a nível municipal
}
\description{
A função extrai os dados de mortalidade infantil e nascimento direto da plataforma
do DATASUS através do pacote datasus e calcula a taxa de mortalidade infantil para
1000 nascimentos a nível municipal.
}
\examples{
\dontrun{
  get_taxa_mortalidade_infantil(ano = "2023")
}
}
