% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_valor_p_hectare_lp.R
\name{get_valor_p_hectare_lp}
\alias{get_valor_p_hectare_lp}
\title{Extrai os dados do Valor da Produção das Lavouras Permanentes por Hectare}
\usage{
get_valor_p_hectare_lp(cod_ibge = "Brasil")
}
\arguments{
\item{cod_ibge}{Um vetor de caracteres com o código IBGE do município ou "Brasil" para todos os municípios ou "Targets" para os municípios alvo.}
}
\value{
Um dataframe
}
\description{
A função extrai  dados Valor da Produção das Lavouras Permanentes por Hectare
do Senso Agro de 2017 através da API do SIDRA IBGE.
}
\examples{
\dontrun{
  data <- get_valor_p_hectare_lp()
}
}
