% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/pega_dados_indicador_desenvolvimento_economico.R
\name{pega_dados_indicador_desenvolvimento_economico}
\alias{pega_dados_indicador_desenvolvimento_economico}
\title{Coleta e organiza os dados da dimensão de Desenvolvimento Econômico}
\usage{
pega_dados_indicador_desenvolvimento_economico(ano = "last")
}
\arguments{
\item{ano_pib}{Ano para coleta do PIB per capita. Padrão é 2021.}
}
\value{
Um data.frame com os indicadores combinados por município
}
\description{
Esta função combina os principais indicadores de desenvolvimento econômico:
- PIB per capita
- Empregos formais por 100 mil habitantes
- IVS 2010 (renda e infraestrutura)
}
\examples{
\dontrun{
  pega_dados_indicador_desenvolvimento_economico(ano_pib = 2021)
}
}
