% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_dados_indicador_educacao.R,
%   R/pega_dados_indicador_gestao_da_qualidade.r
\name{pega_dados_indicador_educacao}
\alias{pega_dados_indicador_educacao}
\title{Coleta completa dos Indicadores de Educação Municipal}
\usage{
pega_dados_indicador_educacao(ano = "2023", download = TRUE)

pega_dados_indicador_educacao(ano = "2023", download = TRUE)
}
\arguments{
\item{ano}{Ano de referência. Padrão: "2023"}

\item{download}{Lógico. Se TRUE, baixa os dados. Padrão: TRUE}

\item{cod_ibge}{Vetor com códigos IBGE. Padrão: target_cities$municipio_codigo}
}
\value{
Uma lista com os dataframes: ideb_iniciais, ideb_finais, tdi, abandono,
matricula_creche, gastos_educacao.

Uma lista com os dataframes: endividamento, liquidez, poupanca_corrente.
}
\description{
Executa o pipeline de extração dos indicadores educacionais nos municípios.

Executa o pipeline de extração dos indicadores de gestão da qualidade nos municípios.
}
