% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_dados_indicador_gestao_da_qualidade.r
\name{pega_dados_indicador_gestao_qualidade}
\alias{pega_dados_indicador_gestao_qualidade}
\title{Coleta completa dos Indicadores de Gestão da Qulidade}
\usage{
pega_dados_indicador_gestao_qualidade(ano = "last", download = TRUE)
}
\arguments{
\item{ano}{Ano de referência. Padrão: "last" ano mais recente.}

\item{download}{Lógico. Se TRUE, baixa os dados. Padrão: TRUE}
}
\value{
Uma lista com os dataframes: endividamento, liquidez, poupanca_corrente.
}
\description{
Executa o pipeline de extração dos indicadores de gestão da qualidade nos municípios.
}
