% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_dados_indicador_infraestrutura.R
\name{pega_dados_indicador_infraestrutura}
\alias{pega_dados_indicador_infraestrutura}
\title{Coleta completa dos Indicadores de Infraestrutura}
\usage{
pega_dados_indicador_infraestrutura(dir = "data_raw", ano = "last")
}
\arguments{
\item{dir}{Diretorio onde os dados necessários serão baixados e lidos.}

\item{ano}{Ano de interesse dos dados. Por padrão usa "last" que refere-se ao
ano mais recente disponível dos dados}
}
\value{
Um dataframe
}
\description{
Executa o pipeline de extração dos indicadores de infraestrutura nos municípios.
}
\examples{
\dontrun{
  dados_indicador_infraestrutura <- pega_dados_indicador_infraestrutura(
    dir = "data_agro",
    ano = "2024"
  )
}
}
