% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_dados_indicador_mobilidade.R
\name{pega_dados_indicador_mobilidade}
\alias{pega_dados_indicador_mobilidade}
\title{Coleta completa dos Indicadores de Mobilidade Urbana}
\usage{
pega_dados_indicador_mobilidade(ano = "22")
}
\arguments{
\item{ano}{Ano dos dados. Padrão: "22"}
}
\value{
Uma lista com os dataframes: internacoes_transito, obitos_transito, deslocamento_sidra, emissao.
}
\description{
Executa o pipeline de extração dos indicadores de mobilidade urbana nos municípios.
}
