% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_dados_indicador_protecao_social.R
\name{pega_dados_indicador_protecao_social}
\alias{pega_dados_indicador_protecao_social}
\title{Coleta completa dos Indicadores de Proteção Social}
\usage{
pega_dados_indicador_protecao_social(
  cod_municipios_ibge = target_cities$municipio_codigo,
  ano = "2023",
  mes = "06",
  dir = "data_raw"
)
}
\arguments{
\item{cod_municipios_ibge}{Vetor com códigos IBGE dos municípios. Padrão: target_cities$municipio_codigo}

\item{ano}{Ano de referência. Padrão: "2023"}

\item{mes}{Mês de referência. Padrão: "06"}

\item{dir}{Diretório dos dados CRAS. Padrão: "data_raw"}
}
\value{
Uma lista com os dataframes: cad_bf, situacao_pobreza, cras.
}
\description{
Executa o pipeline de extração dos indicadores de proteção social nos municípios.
}
