% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_gastos_educacao_from_csv.R
\name{save_gastos_educacao_from_csv}
\alias{save_gastos_educacao_from_csv}
\title{Salva os gastos com educação por município a partir dos arquivos CSV originais}
\usage{
save_gastos_educacao_from_csv(ano = 2022)
}
\arguments{
\item{ano}{Integer. Ano dos dados a serem carregados (entre 2013 e 2024). Default: 2022.}
}
\value{
Dataframe com os municípios e seus respectivos gastos com educação.
}
\description{
Esta função carrega e processa os dados de gastos públicos com educação
a partir de arquivos CSV baixandos manualmente na plataforma da sinconfi e
salvos dentro de <data_raw/finbra_despesas_por_funcao/>, renomeando os arquivos para <finbra_despesas_por_funcao_{ano}>,
filtrando os dados por município e consolidando os valores.
Em seguida, ela salva os arquivos no formato RData dentro de <data/>.
}
\details{
Os dados foram retirados de <https://siconfi.tesouro.gov.br/siconfi/pages/public/consulta_finbra/finbra_list.jsf>
}
\examples{
gastos_educacao <- save_gastos_totais_educacao(ano = 2023)
head(gastos_educacao)

}
