% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_indicador_desenvolvimento_economico.R
\name{unite_indicador_desenvolvimento_economico}
\alias{unite_indicador_desenvolvimento_economico}
\title{Une as tabelas que compõe o indicador de desenvolvimento econômico}
\usage{
unite_indicador_desenvolvimento_economico(
  pib,
  empregos,
  ivs,
  investimento_per_capita,
  recebe_assist,
  produtividade_agricola_lt,
  produtividade_agricola_lp,
  perc_produtores_jovens,
  perc_pessoas_parentesco,
  prop_sistemas_florestais,
  prop_boas_pastagens,
  empresas_abertas
)
}
\arguments{
\item{pib}{Tabela com os dados de PIB Municipais}

\item{empregos}{Tabela com os dados Emprego por 100 mil Habitatnes}

\item{ivs}{Tabela com os dados Índice de Vulnerabilidade Social}

\item{investimento_per_capita}{Tabela com os dados}

\item{recebe_assist}{Tabela com os dados}

\item{produtividade_agricola_lt}{Tabela com os dados de Produtividade Agrícola
por hectare das lavouras temporárias}

\item{produtividade_agricola_lp}{Tabela com os dados de Produtividade Agrícola
por hectare das lavouras permanentes}

\item{perc_produtores_jovens}{Tabela com os dados}

\item{perc_pessoas_parentesco}{Tabela com os dados}

\item{prop_sistemas_florestais}{Tabela com os dados}

\item{prop_boas_pastagens}{Tabela com os dados}

\item{empresas_abertas}{Tabela com os dados de empresas}
}
\value{
Dataframe com todas as variáveis que compõe o indicador de desenvolvimento econômico
}
\description{
Essa função une as tabelas que compões o indicador de desenvolvimento econômico em um único dataframe
}
