% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_indicador_ed.R
\name{unite_indicador_ed}
\alias{unite_indicador_ed}
\title{Une as tabelas que compõe o indicador de educação}
\usage{
unite_indicador_ed(
  ideb_finais,
  ideb_iniciais,
  tdi,
  abandono_escolar,
  matricula_creche,
  gastos_educacao,
  gestao,
  perc_agricultor_ens_bas,
  perc_alimentos_pnae
)
}
\arguments{
\item{ideb_finais}{Tabela com os dados do IDEB}

\item{ideb_iniciais}{Tabela com os dados do IDEB}

\item{tdi}{Tabela com os dados da Taxa de Distorção da Idade por Séries}

\item{abandono_escolar}{Tabela com os dados da Taxa de Abandono Escolar}

\item{matricula_creche}{Tabela com os dados da Taxa Líquida das Matrículas em Creches}

\item{gastos_educacao}{Tabela com os dados de Gastos Per Capita em Educação}

\item{gestao}{Tabela com os dados de Despesas em Educação}

\item{perc_agricultor_ens_bas}{Tabela com os dados do Percentual de Produtores
Agrícolas que concluíram o ensino básico}

\item{perc_alimentos_pnae}{Tabela com os dados do Percentual de Alimentos da
PNAE adquiridos da Agricultura Familiar}
}
\value{
Dataframe com todas as variáveis que compõe o indicador de educação
}
\description{
Essa função une as tabelas que compões o indicador de educação em um único dataframe
}
