% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_indicador_gestao_qualidade.R
\name{unite_indicador_gestao_qualidade}
\alias{unite_indicador_gestao_qualidade}
\title{Une as tabelas que compõe o indicador de saúde}
\usage{
unite_indicador_gestao_qualidade(endividamento, poupanca_corrente, liquidez)
}
\arguments{
\item{endividamento}{Tabela com os dados de Endividamento}

\item{poupanca_corrente}{Tabela com os dados de Poupança Corrente}

\item{liquidez}{Tabela com os dados de Liquidez}
}
\value{
Dataframe com todas as variáveis que compõe o indicador de saúde
}
\description{
Essa função une as tabelas que compões o indicador de saúde em um único dataframe
}
