% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_indicador_infraestrutura.R
\name{unite_indicador_infraestrutura}
\alias{unite_indicador_infraestrutura}
\title{Une as tabelas que compõe o indicador de infraestrutura}
\usage{
unite_indicador_infraestrutura(
  acess_agua_mun,
  acess_esgot_mun,
  vias_pav_mun,
  cob_inte_mun,
  conect_rural_mun
)
}
\arguments{
\item{acess_agua_mun}{Tabela com os dados de Acesso a Distribuição de Agua dos Municípios}

\item{acess_esgot_mun}{Tabela com os dados de Acesso a Sistema de Esgoto dos Municípios}

\item{vias_pav_mun}{Tabela com os dados de Porcentagem de Vias Pavimentadas dos Municípios}

\item{cob_inte_mun}{Tabela com os dados de Cobertura de Internet Móvel dos Municípios}

\item{conect_rural_mun}{Tabela com os dados de Indicador de Conectividade Rural dos Municípios}
}
\value{
Um dataframe
}
\description{
Essa função une as tabelas que compões o indicador de infraestrutura em um único
dataframe
}
