% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_indicador_mobilidade.R
\name{unite_indicador_mobilidade}
\alias{unite_indicador_mobilidade}
\title{Une as tabelas que compõe o indicador de mobilidade}
\usage{
unite_indicador_mobilidade(
  internacoes_transito,
  obitos_transito,
  deslocamento_sidra,
  emissao
)
}
\arguments{
\item{internacoes_transito}{Tabela com os dados de Internações por Sinistro de
Trânsito por 100 mil Habitantes}

\item{obitos_transito}{Tabela com os dados de Óbitos por Sinistro de
Trânsito por 100 mil Habitantes}

\item{deslocamento_sidra}{Tabela com os dados de Tempo de Deslocamento casa/trabalho
superior a uma hora}

\item{emissao}{Tabela com os dados de Emissão por População}
}
\value{
Dataframe com todas as variáveis que compõe o indicador de mobilidade
}
\description{
Essa função une as tabelas que compões o indicador de mobilidade em um único dataframe
}
