% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_indicador_protecao_social.R
\name{unite_indicador_protecao_social}
\alias{unite_indicador_protecao_social}
\title{Une as tabelas que compõe o indicador de proteção social}
\usage{
unite_indicador_protecao_social(
  cad_bf,
  situacao_pobreza,
  cras,
  ivs,
  perc_estab_dirigidos_por_mulheres,
  paridades_camera
)
}
\arguments{
\item{cad_bf}{Tabela contendo os dados da Proporção de Familias que recebem
Bolsa Familia e tem renda de até meio salário Mínimo e a Proporção de Famílias
que recebem até meio salário mínimo e estão cadastradas no CadÚnico}

\item{situacao_pobreza}{Tabela contendo os dados Famílias em situação de pobreza
e estão cadastradas no CadÚnico}

\item{cras}{Tabela contendo os dados de Cras por 100 mil Habitantes}

\item{ivs}{Tabela contendo os dados de Índice de Vulnerabilidade Social}

\item{perc_estab_dirigidos_por_mulheres}{Tabela contendo os dados do percentual
de Estabelecimentos Agrícolas que são dirigidos por Mulheres}
}
\value{
Dataframe com todas as variáveis que compõe o indicador de proteção social
}
\description{
Essa função une as tabelas que compões o indicador de proteção social em um único dataframe
}
