% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constroi_subnutricao_sisvan.R
\name{constroi_subnutricao_sisvan}
\alias{constroi_subnutricao_sisvan}
\title{Controi os dados de Subnutrição - Indicador de Saúde}
\usage{
constroi_subnutricao_sisvan()
}
\value{
Uma dataframe
}
\description{
Essa função constroi os dados de Subnutrição do Sisvan/Ministério da Saúde	 População de todas as
faixas etárias que estão abaixo do peso ideal, a saber:
crianças de 0 – 10 anos (peso muito baixo, peso baixo para a idade),
adolescentes (magreza acentuada para a idade),
adultos, idosos e gestantes (baixo peso).
}
\examples{
\dontrun{
  data <- constroi_subnutricao_sisvan()
}
}
