% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_RAIS.R
\name{download_RAIS}
\alias{download_RAIS}
\title{Download RAIS 2022 Data}
\usage{
download_RAIS(dir_path = "data_raw/")
}
\arguments{
\item{dir_path}{A character string specifying the directory where the file will be saved. Default is "data_raw/".}
}
\value{
No return value. The function downloads the file and saves it locally.
}
\description{
This function downloads the RAIS 2022 data file from the official government website
and saves it in the specified directory.
}
\examples{
download_RAIS("my_data/")
download_RAIS()

}
