% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_dados_indicador_mobilidade.R
\name{get_internacoes_DATASUS}
\alias{get_internacoes_DATASUS}
\title{Extração dos dados de Internações por acidentes de trânsito por munícipio do DATASUS}
\usage{
get_internacoes_DATASUS(ano)
}
\arguments{
\item{ano}{Ano dos dados que serão requisitados. É necessário colocar apenas a década
do ano de interesse no formato string, por exemplo: 2022, ano = "22"}
}
\value{
Uma tabela com nome do município, quantidade de internações, código
do estado segundo IBGE e o ano dos dados
}
\description{
A função extrai os dados de internações por acidente de trânsito a nível municipal
por ano da plataforma do DATASUS. A função retorna uma tabela com nome do município,
quantidade de internações, código do estado segundo IBGE e o ano dos dados
}
\examples{
\dontrun{
  get_internacoes_DATASUS(ano = "09")
}
}
