% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recebe_assist.R
\name{get_recebe_assist}
\alias{get_recebe_assist}
\title{Extrai os dados de Estabelecimentos Agropecuários que recebem Assistência Técnica}
\usage{
get_recebe_assist()
}
\value{
Um dataframe
}
\description{
A função extrai  dados de Estabelecimentos Agropecuários que recebem Assistência Técnica
do Senso Agro de 2017 através da API do SIDRA IBGE.
}
\examples{
\dontrun{
  data <- get_recebe_assist()
}
}
